Display current spectators count on game page
This is basically me feeling out the idea of using polling instead of
web sockets. It feels a bit weird. Going to keep playing around with it.

The intended behavior here is:

- On page load, load the latest game data, including the current
  spectator count
- While the game is active, refresh the game data every 5 seconds
- When the game is dormant, refresh the game data every 2 minutes

That means that when the game is dormant, the current spectator count
(and any other data we display in the future) will be up to two minutes
out of date. We can tune this a bit. I'm hoping to stay on the Heroku
free plan for the backend.