Start tracking GameActions
This starts to sketch out how I'm planning to model the gameplay, as a
linear series of actions that can be inspected at any point to
understand where all the chips are, and why.

The Dealer class is going to be responsible for a lot of that logic. It
represents a dealer sitting at a poker table keeping an eye on everyone.
Is it too cute? Yes?

We're able to drop two columns that are no longer needed, which makes me
feel pretty good about the direction.