Send data about players to frontend
The next step will be to display that, and to add something like:

Add widget for displaying audit log of actions to all humans.

if players.count >= 2
  show "start game" button

when click "start game" button
  - transition game status to playing
  - pick a random player to be the dealer (emit action)
  - make the next two players put in their ante (emit actions)
  - make all the players draw a card (emit actions). Take care when
    displaying these actions to show the card to the right people.

Later:
  - Make the log group items by hand and collapse all but the most
    recent group

Unrelated, but I just realized I'm going to need to figure out how to
handle split pots. That'll be fun.